+++ web/app/components/workflow-app/hooks/index.ts
@@ -6,3 +6,4 @@ export * from './use-workflow-start-run'
 export * from './use-is-chat-mode'
 export * from './use-workflow-refresh-draft'
 export * from './use-fetch-workflow-inspect-vars'
+export * from './use-inspect-vars-crud'
